pyproject.toml
[tool.poetry]
name = "gantt"
version = "0.1.0"
description = ""
authors = ["Ruslan Urazbakhtin <r.urazbahtin@samolet.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
django = "^4.2.1"
djlint = "^1.28.0"
django-jsonform = "^2.17.0"
djangorestframework = "^3.14.0"
django-select2 = "^8.1.2"
django-admin-autocomplete-filter = "^0.7.1"
django-admin-rangefilter = "^0.10.0"
django-filter = "^23.2"
django-bootstrap-datepicker-plus = "^5.0.3"
pillow = "^9.5.0"
django-imagekit = "^4.1.0"
django-extensions = "^3.2.1"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pep8-naming = "^0.13.3"
flake8-django = "^1.3"
flake8-simplify = "^0.20.0"
flake8-bugbear = "^23.5.9"
flake8-comprehensions = "^3.12.0"
flake8-mutable = "^1.2.0"
flake8-expression-complexity = "^0.0.11"
flake8-cognitive-complexity = "^0.1.0"
flake8-gl-codeclimate = "^0.2.1"
dj-database-url = "^2.0.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.13.2"

[tool.black]
line_length = 120
length_sort = true
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120
filter_files = true
skip_glob = ["*/migrations/*", "*/test/*", "*/tests/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
